+++ src/Symfony/Component/Cache/Traits/MemcachedTrait.php
@@ -99,18 +99,42 @@ trait MemcachedTrait
                 if (\is_array($dsn)) {
                     continue;
                 }
-                if (0 !== strpos($dsn, 'memcached://')) {
-                    throw new InvalidArgumentException(sprintf('Invalid Memcached DSN: %s does not start with "memcached://"', $dsn));
+                if (0 !== strpos($dsn, 'memcached:')) {
+                    throw new InvalidArgumentException(sprintf('Invalid Memcached DSN: %s does not start with "memcached:"', $dsn));
                 }
-                $params = preg_replace_callback('#^memcached://(?:([^@]*+)@)?#', function ($m) use (&$username, &$password) {
-                    if (!empty($m[1])) {
-                        list($username, $password) = explode(':', $m[1], 2) + array(1 => null);
+                $params = preg_replace_callback('#^memcached:(//)?(?:([^@]*+)@)?#', function ($m) use (&$username, &$password) {
+                    if (!empty($m[2])) {
+                        list($username, $password) = explode(':', $m[2], 2) + array(1 => null);
                     }
 
-                    return 'file://';
+                    return 'file:'.($m[1] ?? '');
                 }, $dsn);
                 if (false === $params = parse_url($params)) {
                     throw new InvalidArgumentException(sprintf('Invalid Memcached DSN: %s', $dsn));
+                }
+                $query = $hosts = array();
+                if (isset($params['query'])) {
+                    parse_str($params['query'], $query);
+
+                    if (isset($query['host'])) {
+                        if (!\is_array($hosts = $query['host'])) {
+                            throw new InvalidArgumentException(sprintf('Invalid Memcached DSN: %s', $dsn));
+                        }
+                        foreach ($hosts as $host => $weight) {
+                            if (false === $port = strrpos($host, ':')) {
+                                $hosts[$host] = array($host, 11211, (int) $weight);
+                            } else {
+                                $hosts[$host] = array(substr($host, 0, $port), (int) substr($host, 1 + $port), (int) $weight);
+                            }
+                        }
+                        $hosts = array_values($hosts);
+                        unset($query['host']);
+                    }
+                    if ($hosts && !isset($params['host']) && !isset($params['path'])) {
+                        unset($servers[$i]);
+                        $servers = array_merge($servers, $hosts);
+                        continue;
+                    }
                 }
                 if (!isset($params['host']) && !isset($params['path'])) {
                     throw new InvalidArgumentException(sprintf('Invalid Memcached DSN: %s', $dsn));
@@ -124,13 +148,16 @@ trait MemcachedTrait
                     'port' => isset($params['host']) ? 11211 : null,
                     'weight' => 0,
                 );
-                if (isset($params['query'])) {
-                    parse_str($params['query'], $query);
+                if ($query) {
                     $params += $query;
                     $options = $query + $options;
                 }
 
                 $servers[$i] = array($params['host'], $params['port'], $params['weight']);
+
+                if ($hosts) {
+                    $servers = array_merge($servers, $hosts);
+                }
             }
 
             // set client's options